Status finalizado

+++ app/Http/Controllers/StatusController.php
@@ -2,7 +2,6 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Classificacao;
 use App\Models\Status;
 use Illuminate\Http\Request;
 
@@ -19,24 +18,42 @@ class StatusController extends Controller
 
     public function store(Request $request)
     {
+
         $request->validate([
-            "name"=> "required",
+            'name' => 'required|string|max:255', // Exemplo de regra de validação
         ]);
-        $status = Status::create($request->all());
-        return redirect()->route("status.index")->with("success","Created");
+
+        // Cria um novo status com base nos dados recebidos
+        $status = new Status();
+        $status->name = $request->name; // Atribui o nome do status
+        // Outros atributos, se houver
+
+        // Salva o status no banco de dados
+        $status->save();
+
+        // Retorna uma resposta JSON de sucesso
+        return response()->json(['message' => 'Status criado com sucesso', 'status' => $status], 201);
     }
 
-    public function update(Request $request, Status $status)
+    public function update(Request $request, $id)
     {
+        $status = Status::find($id);
         $request->validate([
             "name"=> "required",
         ]);
         $status->update($request->all());
-        redirect()->route("status.index")->with("success","Updated");
+        
+        return response()->json(['message' => 'Status editado com sucesso', 'status' => $status], 201);
+    
     }
 
-    public function destroy(Status $status)
+    public function destroy($id)
     {
-        $status->delete();
+
+        $status = Status::find($id);
+        if($status){
+            $status->delete();
+        }
+        return response()->json(['message' => 'Status excluido com sucesso', 'status' => $status], 201);
     }
 }
